Aligning references, add tests
