+++ src/devices/Common/ClassVisibility.cs
@@ -53,6 +53,10 @@ namespace Iot.Device.Common
     internal partial class NumberHelper
     {
     }
+
+    internal partial struct Window
+    {
+    }
 }
 
 namespace System.Device
