+++ src/devices/Common/Interop/Unix/Libc/Interop.ioctl.macros.cs
@@ -30,6 +30,20 @@ partial class Interop
     private const int _IOC_WRITE = 1;
     private const int _IOC_READ = 2;
 
+    /*
+* Used to create numbers.
+*
+* NOTE: _IOW means userland is writing and kernel is reading. _IOR
+* means userland is reading and kernel is writing.
+#define _IO(type,nr) _IOC(_IOC_NONE,(type),(nr),0)
+#define _IOR(type,nr,size) _IOC(_IOC_READ,(type),(nr),(_IOC_TYPECHECK(size)))
+#define _IOW(type,nr,size) _IOC(_IOC_WRITE,(type),(nr),(_IOC_TYPECHECK(size)))
+#define _IOWR(type,nr,size) _IOC(_IOC_READ|_IOC_WRITE,(type),(nr),(_IOC_TYPECHECK(size)))
+#define _IOR_BAD(type,nr,size) _IOC(_IOC_READ,(type),(nr),sizeof(size))
+#define _IOW_BAD(type,nr,size) _IOC(_IOC_WRITE,(type),(nr),sizeof(size))
+#define _IOWR_BAD(type,nr,size) _IOC(_IOC_READ|_IOC_WRITE,(type),(nr),sizeof(size))
+*/
+
     public static int _IOC(int dir, int type, int nr, int size)
             => ((dir) << _IOC_DIRSHIFT) | ((type) << _IOC_TYPESHIFT) | ((nr) << _IOC_NRSHIFT) | ((size) << _IOC_SIZESHIFT);
