+++ src/devices/Common/Interop/Unix/Libc/Interop.libc.cs
@@ -6,6 +6,7 @@
 #pragma warning disable CS1591 // Public member is not documented - These members are not public in the final package
 
 using System;
+using System.IO;
 using System.Runtime.InteropServices;
 
 partial class Interop
@@ -13,6 +14,7 @@ partial class Interop
     public enum FileOpenFlags
     {
         O_RDONLY = 0x00,
+        O_WRONLY = 0x01,
         O_RDWR = 0x02,
         O_NONBLOCK = 0x800,
         O_SYNC = 0x101000
@@ -38,7 +40,13 @@ partial class Interop
     public static extern int ioctl(int fd, int request, IntPtr argp);
 
     [DllImport(LibcLibrary, SetLastError = true)]
+    public static extern int ioctl(int fd, int request, int intArg);
+
+    [DllImport(LibcLibrary, SetLastError = true)]
     public static extern int open([MarshalAs(UnmanagedType.LPStr)] string pathname, FileOpenFlags flags);
+
+    [DllImport(LibcLibrary)]
+    public static extern int write(int fd, IntPtr data, int length);
 
     [DllImport(LibcLibrary)]
     public static extern int close(int fd);
@@ -48,4 +56,22 @@ partial class Interop
 
     [DllImport(LibcLibrary)]
     public static extern int munmap(IntPtr addr, int length);
+
+    public static void ioctlv(int fd, int request, IntPtr argp)
+    {
+        int result = ioctl(fd, request, argp);
+        if (result != 0)
+        {
+            throw new IOException($"IOCTL request {request} failed: {result}");
+        }
+    }
+
+    public static void ioctlv(int fd, int request, int arg)
+    {
+        int result = ioctl(fd, request, arg);
+        if (result != 0)
+        {
+            throw new IOException($"IOCTL request {request} failed: {result}");
+        }
+    }
 }
