+++ src/devices/Common/Interop/Windows/WinUser.cs
@@ -9,6 +9,87 @@ using System.Runtime.InteropServices;
 
 partial class Interop
 {
+    public enum SystemMetric
+    {
+        SM_CXSCREEN = 0,
+        SM_CYSCREEN = 1,
+        SM_CXVSCROLL = 2,
+        SM_CYHSCROLL = 3,
+        SM_CYCAPTION = 4,
+        SM_CXBORDER = 5,
+        SM_CYBORDER = 6,
+        SM_CXFIXEDFRAME = 7,
+        SM_CYFIXEDFRAME = 8,
+        SM_CYVTHUMB = 9,
+        SM_CXHTHUMB = 10,
+        SM_CXICON = 11,
+        SM_CYICON = 12,
+        SM_CXCURSOR = 13,
+        SM_CYCURSOR = 14,
+        SM_CYMENU = 15,
+        SM_CYKANJIWINDOW = 18,
+        SM_MOUSEPRESENT = 19,
+        SM_CYVSCROLL = 20,
+        SM_CXHSCROLL = 21,
+        SM_DEBUG = 22,
+        SM_SWAPBUTTON = 23,
+        SM_CXMIN = 28,
+        SM_CYMIN = 29,
+        SM_CXSIZE = 30,
+        SM_CYSIZE = 31,
+        SM_CXFRAME = 32,
+        SM_CYFRAME = 33,
+        SM_CXMINTRACK = 34,
+        SM_CYMINTRACK = 35,
+        SM_CXDOUBLECLK = 36,
+        SM_CYDOUBLECLK = 37,
+        SM_CXICONSPACING = 38,
+        SM_CYICONSPACING = 39,
+        SM_MENUDROPALIGNMENT = 40,
+        SM_PENWINDOWS = 41,
+        SM_DBCSENABLED = 42,
+        SM_CMOUSEBUTTONS = 43,
+        SM_SECURE = 44,
+        SM_CXEDGE = 45,
+        SM_CYEDGE = 46,
+        SM_CXMINSPACING = 47,
+        SM_CYMINSPACING = 48,
+        SM_CXSMICON = 49,
+        SM_CYSMICON = 50,
+        SM_CYSMCAPTION = 51,
+        SM_CXSMSIZE = 52,
+        SM_CYSMSIZE = 53,
+        SM_CXMENUSIZE = 54,
+        SM_CYMENUSIZE = 55,
+        SM_ARRANGE = 56,
+        SM_CXMINIMIZED = 57,
+        SM_CYMINIMIZED = 58,
+        SM_CXMAXTRACK = 59,
+        SM_CYMAXTRACK = 60,
+        SM_CXMAXIMIZED = 61,
+        SM_CYMAXIMIZED = 62,
+        SM_NETWORK = 63,
+        SM_CLEANBOOT = 67,
+        SM_CXDRAG = 68,
+        SM_CYDRAG = 69,
+        SM_SHOWSOUNDS = 70,
+        SM_CXMENUCHECK = 71,
+        SM_CYMENUCHECK = 72,
+        SM_MIDEASTENABLED = 74,
+        SM_MOUSEWHEELPRESENT = 75,
+        SM_XVIRTUALSCREEN = 76,
+        SM_YVIRTUALSCREEN = 77,
+        SM_CXVIRTUALSCREEN = 78,
+        SM_CYVIRTUALSCREEN = 79,
+        SM_CMONITORS = 80,
+        SM_SAMEDISPLAYFORMAT = 81,
+        SM_CYFOCUSBORDER = 84,
+        SM_CXFOCUSBORDER = 83,
+        SM_REMOTESESSION = 0x1000,
+        SM_CXSIZEFRAME = SM_CXFRAME,
+        SM_CYSIZEFRAME = SM_CYFRAME
+    }
+
     public static int VK_NUMLOCK = 0x90;
     public static int VK_SCROLL = 0x91;
     public static int VK_CAPITAL = 0x14;
@@ -27,4 +108,86 @@ partial class Interop
 
     [DllImport("User32.dll", SetLastError = true)]
     public static extern short GetAsyncKeyState(int vKey);
+
+    [DllImport("User32.dll", ExactSpelling = true)]
+    public static extern int GetSystemMetrics(SystemMetric nIndex);
+
+    [Flags]
+    public enum MouseEventFlags
+    {
+        LeftDown = 0x00000002,
+        LeftUp = 0x00000004,
+        MiddleDown = 0x00000020,
+        MiddleUp = 0x00000040,
+        Move = 0x00000001,
+        Absolute = 0x00008000,
+        RightDown = 0x00000008,
+        RightUp = 0x00000010
+    }
+
+    [StructLayout(LayoutKind.Sequential)]
+    public struct MousePoint
+    {
+        public int X;
+        public int Y;
+
+        public MousePoint(int x, int y)
+        {
+            X = x;
+            Y = y;
+        }
+    }
+
+    [DllImport("user32.dll", EntryPoint = "SetCursorPos")]
+    [return: MarshalAs(UnmanagedType.Bool)]
+    private static extern bool SetCursorPos(int x, int y);
+
+    [DllImport("user32.dll")]
+    [return: MarshalAs(UnmanagedType.Bool)]
+    private static extern bool GetCursorPos(out MousePoint lpMousePoint);
+
+    [DllImport("user32.dll")]
+    private static extern void mouse_event(int dwFlags, int dx, int dy, int dwData, int dwExtraInfo);
+
+    public static void SetCursorPosition(int x, int y)
+    {
+        SetCursorPos(x, y);
+    }
+
+    public static void SetCursorPosition(MousePoint point)
+    {
+        SetCursorPos(point.X, point.Y);
+    }
+
+    public static MousePoint GetCursorPosition()
+    {
+        MousePoint currentMousePoint;
+        var gotPoint = GetCursorPos(out currentMousePoint);
+        if (!gotPoint)
+        {
+            currentMousePoint = new MousePoint(0, 0);
+        }
+
+        return currentMousePoint;
+    }
+
+    public static void MouseEvent(MouseEventFlags value)
+    {
+        MousePoint position = GetCursorPosition();
+
+        mouse_event((int)value,
+                position.X,
+                position.Y,
+                0,
+                0);
+    }
+
+    public static void MouseEvent(MousePoint position, MouseEventFlags value)
+    {
+        mouse_event((int)value,
+            position.X,
+            position.Y,
+            0,
+            0);
+    }
 }
