+++ src/devices/Gui/Gui.csproj
@@ -5,10 +5,14 @@
     <!--Disabling default items so samples source won't get build by the main library-->
     <EnableDefaultItems>false</EnableDefaultItems>
     <LangVersion>9</LangVersion>
+    <RootNamespace>Iot.Device.Gui</RootNamespace>
+    <AllowUnsafeBlocks>True</AllowUnsafeBlocks>
   </PropertyGroup>
 
   <ItemGroup>
     <Compile Include="*.cs" />
+    <Compile Include="Interop\Interop.uinput.cs" />
+    <Compile Include="Interop\Interop.X11.cs" />
     <None Include="README.md" />
   </ItemGroup>
