+++ src/devices/Gui/Gui.sln
@@ -1,17 +1,21 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.26124.0
+# Visual Studio Version 17
+VisualStudioVersion = 17.7.33808.371
 MinimumVisualStudioVersion = 15.0.26124.0
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Gui", "Gui.csproj", "{B82C190A-642B-465B-BD3F-DB56FFF22253}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Gui", "Gui.csproj", "{B82C190A-642B-465B-BD3F-DB56FFF22253}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "tests", "tests", "{03107CFF-FE00-467D-9B38-C00F7E12A7FE}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Gui.Tests", "tests\Gui.Tests.csproj", "{2B0F5D91-F38B-427D-84DB-5656D9661E1C}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Gui.Tests", "tests\Gui.Tests.csproj", "{2B0F5D91-F38B-427D-84DB-5656D9661E1C}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "samples", "samples", "{6A4DE7B1-03F3-4EE0-BF73-A0BAEF88BA2B}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Gui.Samples", "samples\Gui.Samples.csproj", "{3CFA13D6-1D29-4C87-B0C1-01A6901A50EF}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Gui.Sample", "samples\Gui.Sample.csproj", "{3CFA13D6-1D29-4C87-B0C1-01A6901A50EF}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Common", "..\Common\Common.csproj", "{52BD85AE-307A-4EC2-9CE5-E0D4D2873764}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SkiaSharpAdapter", "..\SkiaSharpAdapter\SkiaSharpAdapter.csproj", "{9B078F42-5226-44A3-974E-1BE23511557A}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -22,10 +26,19 @@ Global
 		Release|x64 = Release|x64
 		Release|x86 = Release|x86
 	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{B82C190A-642B-465B-BD3F-DB56FFF22253}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B82C190A-642B-465B-BD3F-DB56FFF22253}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B82C190A-642B-465B-BD3F-DB56FFF22253}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{B82C190A-642B-465B-BD3F-DB56FFF22253}.Debug|x64.Build.0 = Debug|Any CPU
+		{B82C190A-642B-465B-BD3F-DB56FFF22253}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{B82C190A-642B-465B-BD3F-DB56FFF22253}.Debug|x86.Build.0 = Debug|Any CPU
+		{B82C190A-642B-465B-BD3F-DB56FFF22253}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B82C190A-642B-465B-BD3F-DB56FFF22253}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B82C190A-642B-465B-BD3F-DB56FFF22253}.Release|x64.ActiveCfg = Release|Any CPU
+		{B82C190A-642B-465B-BD3F-DB56FFF22253}.Release|x64.Build.0 = Release|Any CPU
+		{B82C190A-642B-465B-BD3F-DB56FFF22253}.Release|x86.ActiveCfg = Release|Any CPU
+		{B82C190A-642B-465B-BD3F-DB56FFF22253}.Release|x86.Build.0 = Release|Any CPU
 		{2B0F5D91-F38B-427D-84DB-5656D9661E1C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{2B0F5D91-F38B-427D-84DB-5656D9661E1C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{2B0F5D91-F38B-427D-84DB-5656D9661E1C}.Debug|x64.ActiveCfg = Debug|Any CPU
@@ -50,21 +63,39 @@ Global
 		{3CFA13D6-1D29-4C87-B0C1-01A6901A50EF}.Release|x64.Build.0 = Release|Any CPU
 		{3CFA13D6-1D29-4C87-B0C1-01A6901A50EF}.Release|x86.ActiveCfg = Release|Any CPU
 		{3CFA13D6-1D29-4C87-B0C1-01A6901A50EF}.Release|x86.Build.0 = Release|Any CPU
-		{B82C190A-642B-465B-BD3F-DB56FFF22253}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{B82C190A-642B-465B-BD3F-DB56FFF22253}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{B82C190A-642B-465B-BD3F-DB56FFF22253}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{B82C190A-642B-465B-BD3F-DB56FFF22253}.Debug|x64.Build.0 = Debug|Any CPU
-		{B82C190A-642B-465B-BD3F-DB56FFF22253}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{B82C190A-642B-465B-BD3F-DB56FFF22253}.Debug|x86.Build.0 = Debug|Any CPU
-		{B82C190A-642B-465B-BD3F-DB56FFF22253}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{B82C190A-642B-465B-BD3F-DB56FFF22253}.Release|Any CPU.Build.0 = Release|Any CPU
-		{B82C190A-642B-465B-BD3F-DB56FFF22253}.Release|x64.ActiveCfg = Release|Any CPU
-		{B82C190A-642B-465B-BD3F-DB56FFF22253}.Release|x64.Build.0 = Release|Any CPU
-		{B82C190A-642B-465B-BD3F-DB56FFF22253}.Release|x86.ActiveCfg = Release|Any CPU
-		{B82C190A-642B-465B-BD3F-DB56FFF22253}.Release|x86.Build.0 = Release|Any CPU
+		{52BD85AE-307A-4EC2-9CE5-E0D4D2873764}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{52BD85AE-307A-4EC2-9CE5-E0D4D2873764}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{52BD85AE-307A-4EC2-9CE5-E0D4D2873764}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{52BD85AE-307A-4EC2-9CE5-E0D4D2873764}.Debug|x64.Build.0 = Debug|Any CPU
+		{52BD85AE-307A-4EC2-9CE5-E0D4D2873764}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{52BD85AE-307A-4EC2-9CE5-E0D4D2873764}.Debug|x86.Build.0 = Debug|Any CPU
+		{52BD85AE-307A-4EC2-9CE5-E0D4D2873764}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{52BD85AE-307A-4EC2-9CE5-E0D4D2873764}.Release|Any CPU.Build.0 = Release|Any CPU
+		{52BD85AE-307A-4EC2-9CE5-E0D4D2873764}.Release|x64.ActiveCfg = Release|Any CPU
+		{52BD85AE-307A-4EC2-9CE5-E0D4D2873764}.Release|x64.Build.0 = Release|Any CPU
+		{52BD85AE-307A-4EC2-9CE5-E0D4D2873764}.Release|x86.ActiveCfg = Release|Any CPU
+		{52BD85AE-307A-4EC2-9CE5-E0D4D2873764}.Release|x86.Build.0 = Release|Any CPU
+		{9B078F42-5226-44A3-974E-1BE23511557A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{9B078F42-5226-44A3-974E-1BE23511557A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{9B078F42-5226-44A3-974E-1BE23511557A}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{9B078F42-5226-44A3-974E-1BE23511557A}.Debug|x64.Build.0 = Debug|Any CPU
+		{9B078F42-5226-44A3-974E-1BE23511557A}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{9B078F42-5226-44A3-974E-1BE23511557A}.Debug|x86.Build.0 = Debug|Any CPU
+		{9B078F42-5226-44A3-974E-1BE23511557A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{9B078F42-5226-44A3-974E-1BE23511557A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{9B078F42-5226-44A3-974E-1BE23511557A}.Release|x64.ActiveCfg = Release|Any CPU
+		{9B078F42-5226-44A3-974E-1BE23511557A}.Release|x64.Build.0 = Release|Any CPU
+		{9B078F42-5226-44A3-974E-1BE23511557A}.Release|x86.ActiveCfg = Release|Any CPU
+		{9B078F42-5226-44A3-974E-1BE23511557A}.Release|x86.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{2B0F5D91-F38B-427D-84DB-5656D9661E1C} = {03107CFF-FE00-467D-9B38-C00F7E12A7FE}
 		{3CFA13D6-1D29-4C87-B0C1-01A6901A50EF} = {6A4DE7B1-03F3-4EE0-BF73-A0BAEF88BA2B}
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {8316557A-2709-4CA2-A5AF-5C57E4B08030}
 	EndGlobalSection
 EndGlobal
