+++ src/devices/Gui/tests/GuiTests..cs
@@ -1,16 +1,33 @@
-// Licensed to the .NET Foundation under one or more agreements.
+﻿// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
+using System;
+using System.Runtime.InteropServices;
 using Xunit;
 
-namespace Iot.Device._DeviceBinding.Tests
+namespace Iot.Device.Gui.Tests
 {
-    public class _DeviceBindingTests
+    public class ScreenCaptureTests
     {
         [Fact]
-        public void Test1()
+        public void TestWindows()
         {
-            // NOTE: THIS TEST PROJECT IS NOT REQUIRED AND CAN BE DELETED IF NOT USED.
+            var screenCapture = new ScreenCapture();
+            if (OperatingSystem.IsWindows() || OperatingSystem.IsLinux())
+            {
+                var size = screenCapture.ScreenSize();
+                Assert.True(size.Width > 0);
+                Assert.True(size.Height > 0);
+                var img = screenCapture.GetScreenContents();
+                Assert.NotNull(img);
+                Assert.Equal(size.Width, img!.Width);
+                Assert.Equal(size.Height, img.Height);
+            }
+            else
+            {
+                Assert.Throws<PlatformNotSupportedException>(() => screenCapture.GetScreenContents());
+                Assert.Throws<PlatformNotSupportedException>(() => screenCapture.ScreenSize());
+            }
         }
     }
 }
